Improve grouping of domain models into files
