--- src/main/kotlin/dev/pankowski/garcon/domain/SlackReposter.kt
@@ -1,6 +1,0 @@
-package dev.pankowski.garcon.domain
-
-interface SlackReposter {
-
-  fun repost(post: Post, pageName: PageName)
-}
