+++ src/main/kotlin/dev/pankowski/garcon/infrastructure/RestTemplateSlack.kt
@@ -8,8 +8,7 @@ import org.springframework.stereotype.Component
 import org.springframework.web.client.RestTemplate
 
 @Component
-class RestTemplateSlackReposter(private val slackConfig: SlackConfig, restTemplateBuilder: RestTemplateBuilder) :
-  SlackReposter {
+class RestTemplateSlack(private val slackConfig: SlackConfig, restTemplateBuilder: RestTemplateBuilder) : Slack {
 
   private val log = getLogger(javaClass)
